Templating engine complete, however bug with md loading
The templating engine is complete, however there is a bug where the xhttp request in the md to html function will only find the home.header.md file, not the info.header.md or test.header.md even thought they are all in the same directory. 

The templating engine code also needs to be commented.